requirement.txt
backcall==0.1.0
decorator==4.3.0
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.4
django-daterange-filter==1.3.0
django-debug-toolbar==1.11
django-grappelli==2.12.1
django-import-export==1.1.0
et-xmlfile==1.0.1
ipython==7.2.0
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.13.2
odfpy==1.4.0
openpyxl==2.5.12
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2018.7
PyYAML==3.13
six==1.12.0
sqlparse==0.2.4
tablib==0.12.1
traitlets==4.3.2
unicodecsv==0.14.1
wcwidth==0.1.7
xlrd==1.2.0
xlwt==1.3.0
